+++ metadata_mapper/mappers/marc/ucb_tind_mapper.py
@@ -6,44 +6,54 @@ from pymarc import parse_xml_to_array
 from lxml import etree
 from io import StringIO
 
-from typing import Callable
+from ..mapper import Validator
+
 from collections import OrderedDict
 
 import re
+
 
 class UcbTindRecord(MarcRecord):
     def UCLDC_map(self):
         return {
             "calisphere-id": self.legacy_couch_db_id.split("--")[1],
             "_id": self.get_marc_data_fields(["901"], ["a"]),
-            "isShownAt": self.get_marc_data_fields(["856"], ["u"]),
-            "isShownBy": self.get_marc_data_fields(["856"], ["u"]),
+            "isShownAt": self.map_is_shown_at,
+            "isShownBy": self.map_is_shown_by,
+            "alt_title": self.get_marc_data_fields(["246"], ["6"],
+                                                   exclude_subfields=True),
             "language": self.get_marc_data_fields(["041"], ["a"]),
             "date": self.get_marc_data_fields(["260"], ["c"]),
             "publisher": self.get_marc_data_fields(["260"], ["a", "b"]),
-            "format": self.map_format,
+            "format": self.get_marc_data_fields(["655"], ["2"],
+                                                exclude_subfields=True),
             "extent": self.map_extent,
-            "identifier": self.get_marc_data_fields(["020", "022", "024", "901"],
+            "identifier": self.get_marc_data_fields(["024", "901", "035"],
                                                     ["a"]),
-            "creator": self.get_marc_data_fields(["100", "110", "111"]),
+            "contributor": self.get_marc_data_fields(["100", "110", "111"]),
+            "creator": self.get_marc_data_fields(["700"], ["a"]),
             "relation": self.map_relation,
+            "provenance": self.get_marc_data_fields(["541"], ["a"]),
             "description": self.map_description,
             "rights": self.get_marc_data_fields(["506", "540"]),
             "temporal": self.get_marc_data_fields(["648"]),
-            "contributor": self.get_marc_data_fields(["700", "710", "711", "720"]),
             "title": self.map_title,
             "spatial": self.map_spatial,
             "spec_type": self.map_spec_type,
             "subject": self.map_subject,
-            "type": self.map_type,
+            "type": self.get_marc_data_fields(["336"])
         }
 
-    def map_type(self):
-        value = []
-        for types in self.get_matching_types():
-            value.append(types[1])
+    def map_is_shown_at(self):
+        field_001 = self.get_marc_control_field("001")
+        if field_001:
+            return "https://digicoll.lib.berkeley.edu/record/" + field_001
 
-        return value
+    def map_is_shown_by(self):
+        field_001 = self.get_marc_control_field("001")
+        if field_001:
+            return ("https://digicoll.lib.berkeley.edu/nanna/thumbnail/v2/" +
+                    field_001 + "?redirect=1")
 
     def map_spec_type(self):
         value = []
@@ -87,58 +97,9 @@ class UcbTindRecord(MarcRecord):
 
         return f651 + self.get_marc_data_fields(self.get_metadata_fields(), ["z"])
 
-    def map_format(self) -> list:
-        f3xx = self.get_marc_data_fields(["337", "338", "340"], ["a"]),
-
-        return f3xx + self.get_marc_data_fields(self.get_metadata_fields(), ["v"])
-
     def map_subject(self) -> list:
-
-        def get_delimiters(tag, code):
-            """
-            Returns the appropriate delimiter(s) based on the tag and code
-            """
-            if tag == "658":
-                if code == "b":
-                    return [":"]
-                elif code == "c":
-                    return [" [", "]"]
-                elif code == "d":
-                    return ["--"]
-            elif ((tag == "653") or (int(tag) in range(690, 700)) or
-                  (code == "b" and
-                   tag in ("654", "655")) or (code in ("v", "x", "y", "z"))):
-                return ["--"]
-            elif (tag == "610"):
-                if code == "b":
-                    return [" "]
-                else:
-                    return ["--"]
-            elif code == "d":
-                return [", "]
-
-            return [". "]
-
-        def split_subject(value, tag, code):
-            delimiters = get_delimiters(tag, code)
-
-            # Skip codes that are numeric
-            if not code or code.isdigit():
-                return
-
-            value = value.rstrip(", ")
-
-            if value:
-                for delimiter in delimiters:
-                    values = [delimiter.join(values)]
-                    if delimiter != delimiters[-1]:
-                        # Append an empty value for subsequent joins
-                        values.append("")
-
-            return values
-
-        return self.get_marc_data_fields(self.get_metadata_fields(),
-                                         process_value=split_subject)
+        fields = self.get_metadata_fields()
+        return [{"name": s} for s in self.get_marc_data_fields(fields, ["2"], exclude_subfields=True)]
 
     def map_temporal(self) -> list:
         f648 = self.get_marc_data_fields(["648"])
@@ -169,8 +130,8 @@ class UcbTindRecord(MarcRecord):
 
         :return: A list of extent values.
         """
-        return self.get_marc_data_fields(["300"]) + self.get_marc_data_fields(["340"],
-                                                                              ["b"])
+        return [", ".join(self.get_marc_data_fields(["300"]) + self.get_marc_data_fields(["340"],
+                                                                              ["b"]))]
 
     def map_title(self) -> list:
         # 245, all subfields except c
@@ -184,126 +145,6 @@ class UcbTindRecord(MarcRecord):
 
         return f245 + f242 + f240
 
-    def get_marc_control_field(self, field_tag: str, index: int = None) -> list:
-        """
-        Get MARC control field. Returns an empty string if:
-            * Control field isn't set
-            * No value exists at the requested index
-        Otherwise it returns a value
-
-        TODO: maybe need to accept slices in addition to ints for the index
-
-        :param field_tag: Field tag to retrieve.
-        :param index: A specific index to fetch
-        :return: List of values for the control fields.
-        """
-
-        # Don't let any data tags sneak in! They have subfields.
-        data_field_tag = field_tag if field_tag.isnumeric() and int(
-            field_tag) < 100 else ""
-
-        values = [v[0].value() for (k, v)
-                  in self.get_marc_tag_value_map([data_field_tag]).items()
-                  if len(v) > 0]
-
-        if not values:
-            return ""
-
-        value = values[0]
-
-        if index and len(value) > index + 1:
-            return value[index]
-
-        if index:
-            return ""
-
-        return value
-
-    def get_marc_data_fields(self, field_tags: list, subfield_codes=[],
-                             **kwargs) -> list:
-        """
-        Get the values of specified subfields from given MARC fields. This allows control fields too.
-
-        Set the `exclude_subfields` kwarg to exclude the specified subfield_codes.
-
-        Set the `process_value` kwarg to pass the value through your own code to
-        do transformations based on the field tag, code and value. See `map_subject` for
-        an example.
-
-        :param field_tags: A list of MARC fields.
-        :param subfield_codes: A list of subfield codes to filter the values. If empty,
-                               all subfields will be included.
-        :return: A list of values of the specified subfields.
-        """
-        def subfield_matches(check_code: str, subfield_codes: list,
-                             exclude_subfields: bool) -> bool:
-            """
-            :param check_code: The code to check against the subfield codes.
-            :param subfield_codes: A list of subfield codes to include / exclude
-            :param exclude_subfields: A boolean value indicating whether to exclude the
-                                      specified subfield codes.
-            :return: A boolean value indicating whether the check_code is included or
-                    excluded based on the subfield_codes and exclude_subfields parameters.
-            """
-            if not subfield_codes:
-                return True
-            if exclude_subfields:
-                return check_code not in subfield_codes
-            else:
-                return check_code in subfield_codes
-
-        if "process_value" in kwargs and isinstance(kwargs["process_value"], Callable):
-            process_value = kwargs["process_value"]
-        else:
-            process_value = None
-
-        exclude_subfields = "exclude_subfields" in kwargs and kwargs[
-            "exclude_subfields"]
-
-        value_list = [process_value(value, field_tag, subfield[0])
-                      if process_value else value
-                      for (field_tag, matching_fields) in
-                      self.get_marc_tag_value_map(field_tags).items()
-                      for matching_field in matching_fields
-                      for subfield in list(matching_field.subfields_as_dict().items())
-                      for value in subfield[1]
-                      if
-                      subfield_matches(subfield[0], subfield_codes, exclude_subfields)]
-
-        return value_list if isinstance(value_list, list) else []
-
-    def get_marc_tag_value_map(self, field_tags: list) -> dict:
-        """
-        Get the specified MARC fields from the source_metadata, mapping by field tag
-
-        :param field_tags: List of MARC fields to retrieve.
-        :return: List of MARC fields from the source_metadata.
-        """
-        return {field_tag: self.source_metadata.get("marc").get_fields(field_tag) for
-                field_tag in field_tags}
-
-    def get_marc_leader(self, leader_key: str):
-        """
-        Retrieve the value of specified leader key from the MARC metadata.
-
-        Couple things:
-            * We're not accommodating passing a slice, which pymarc can handle should it be necessary
-            * Both
-
-        :param leader_key: The key of the leader field to retrieve.
-        :type leader_key: str
-        :return: The value of the specified leader key.
-        :rtype: str or None
-        """
-        leader = self.source_metadata.get("marc").leader
-
-        if str(leader_key).isnumeric():
-            return leader[int(leader_key)]
-
-        if hasattr(leader, leader_key):
-            return leader.getattr(leader_key, "")
-
-        return ""
 
     def get_types(self):
         """
@@ -369,8 +210,29 @@ class UcbTindRecord(MarcRecord):
         }
 
 
+class UcbTindValidator(Validator):
+
+    def setup(self):
+        self.add_validatable_fields([
+            {
+                "field": "is_shown_by",
+                "validations": [
+                    Validator.str_match_ignore_url_protocol,
+                    Validator.verify_type(str)
+                ]
+            },
+            {
+                "field": "is_shown_at",
+                "validations": [
+                    Validator.str_match_ignore_url_protocol,
+                    Validator.verify_type(str)
+                ]
+            }
+        ])
+
 class UcbTindVernacular(OaiVernacular):
     record_cls = UcbTindRecord
+    validator = UcbTindValidator
 
     def _process_record(self, record_element: list, request_url: str) -> UcbTindRecord:
         """
